Move date formatter into utility script
