+++ utility.py
@@ -5,6 +5,7 @@ import time
 import contextlib
 from http.client import HTTPConnection # py3
 import requests
+import datetime
 import json
 import xml.etree.ElementTree as ET
 
@@ -126,3 +127,22 @@ def get_xml_from_url(url, headers, cache_file_name, ttl):
             response_data = file.read()
     response_xml = ET.fromstring(response_data)
     return response_xml
+
+def get_formatted_date(dt, include_time=True):
+    today = datetime.datetime.today()
+    tomorrow = today + datetime.timedelta(days=1)
+    next_week = today + datetime.timedelta(days=7)
+    formatter_day = "%a %b %-d"
+    formatter_time = ""
+
+    if include_time:
+        formatter_time = ", %-I:%M %p"
+
+    if dt.date() == today.date():
+        formatter_day = "Today"
+    elif dt.date() == tomorrow.date():
+        formatter_day = "Tomorrow"
+    elif dt.date() < next_week.date():
+        formatter_day = "%A"
+    return dt.strftime(formatter_day + formatter_time)
+
